+++ hw5/hw5.cpp
@@ -181,12 +181,12 @@ void display() {
       }
 
     //  Translate intensity to color vectors
-    float Ambient[]   = {0.01*ambient ,0.01*ambient ,0.01*ambient ,1.0};
-    float Diffuse[]   = {0.01*diffuse ,0.01*diffuse ,0.01*diffuse ,1.0};
-    float Specular[]  = {0.01*specular,0.01*specular,0.01*specular,1.0};
+     float Ambient[]   = {(float) 0.01*ambient ,(float) 0.01*ambient ,(float) 0.01*ambient , (float)1.0};
+     float Diffuse[]   = {(float) 0.01*diffuse ,(float) 0.01*diffuse , (float) 0.01*diffuse , (float) 1.0};
+     float Specular[]  = {(float) 0.01*specular, (float) 0.01*specular, (float)0.01*specular, (float) 1.0};
 
     //  Light position
-    float Position[]  = {distance*Cos(zh),ylight,distance*Sin(zh),1.0};
+     float Position[]  = {(float)distance*(float)Cos(zh),(float)ylight,(float)distance*(float)Sin(zh),(float)1.0};
     //  Draw light position as ball (still no lighting here)
     glColor3f(1,1,1);
     Ball b = Ball(Position[0],Position[1],Position[2] , 0.1);
